requirements.txt
attrs
django
django-extended-choices
pytest
pytest-django
psycopg2-binary
ruff